Extract email composer capability
