+++ app/apps/email/components/EmailClient.tsx
@@ -1,38 +1,46 @@
 'use client';
 
-import { useCallback, useEffect, useMemo, useRef, useState, type ChangeEvent, type KeyboardEvent } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import {
   Check,
   ChevronDown,
   ChevronLeft,
   ChevronRight,
-  FileText,
   Folder,
   Inbox,
   Loader2,
-  Mail,
   MailWarning,
   PanelLeftClose,
   PanelLeftOpen,
   PenLine,
-  Plus,
   RefreshCw,
   Search,
-  Sparkles,
   Star,
   Settings,
-  Wrench,
-  X,
 } from 'lucide-react';
 import { useLocale, useTranslations } from 'next-intl';
 import { useSearchParams } from 'next/navigation';
 
-import { EmailAttachmentPanel } from '@/app/apps/email/components/EmailAttachmentPanel';
-import { EmailMessageBody, EmailMessageRowActions, EmailMessageViewer } from '@/app/apps/email/components/EmailMessageReader';
+import { EmailComposeDialog } from '@/app/apps/email/components/EmailComposeDialog';
+import { EmailMessageRowActions, EmailMessageViewer } from '@/app/apps/email/components/EmailMessageReader';
 import { EmailReviewCenter } from '@/app/apps/email/components/EmailReviewCenter';
 import { EmailPaneResizeHandle, useEmailWorkspaceLayout } from '@/app/apps/email/components/EmailWorkspaceLayout';
-import { extractEmailAddressForCompose, formatDate, formatRecipients } from '@/app/apps/email/components/email-client-format';
+import {
+  composeRecipientText,
+  extractRecipientEmailsForCompose,
+  forwardSubjectForCompose,
+  normalizeAgentUsedContext,
+  pruneUnreferencedInlineEmailAttachments,
+  replySubjectForCompose,
+  splitRecipientInput,
+  uniqueComposeRecipients,
+} from '@/app/apps/email/components/email-compose-utils';
+import { extractEmailAddressForCompose, formatDate } from '@/app/apps/email/components/email-client-format';
 import type {
+  EmailComposeAgentToolEvent,
+  EmailComposeDialogLabels,
+  EmailComposeDraft,
+  EmailComposeMode,
   EmailFolder,
   EmailMessageActionName,
   EmailMessageContextMenuPosition,
@@ -40,20 +48,12 @@ import type {
   EmailMessageListActionName,
   EmailMessageListActionState,
   EmailMessageSummary,
-  EmailMessageViewerLabels,
   EmailOutboxDraft,
   WorkspaceInboxCase,
 } from '@/app/apps/email/components/email-client-types';
 import { useSetEmailChatContext } from '@/app/apps/email/context/email-chat-context';
 import { buildEmailPageChatContext } from '@/app/apps/email/context/email-route-chat-context';
-import {
-  ComposerReferencePicker,
-  type ComposerReferencePickerItem,
-} from '@/app/components/canvas-agent-chat/ComposerReferencePicker';
-import type { FilePickerFile } from '@/app/components/canvas-agent-chat/ChatComposer';
 import { EmailAccountsCard } from '@/app/components/settings/IntegrationsSettingsClient';
-import { findActiveComposerReference, replaceComposerReference, type ComposerReferenceMatch } from '@/app/lib/chat/composer-references';
-import type { EmailAttachmentDraft } from '@/app/lib/email/attachment-types';
 import { plainTextToEmailHtml } from '@/app/lib/email/html-conversion';
 import {
   composeEmailEditorBodyValues,
@@ -61,14 +61,11 @@ import {
   sanitizeEmailEditorHtml,
 } from '@/app/lib/email/html-editor-content';
 import { emailMessageContentRevision, emailMessageListScopeKey } from '@/app/lib/email/reader-refresh';
-import { getFileIconComponent } from '@/app/lib/files/file-icons';
-import { listWorkspaceFileReferences } from '@/app/lib/files/client';
-import { getToolDisplayInfo } from '@/app/lib/pi/tool-display';
 import type { NotebookEmailContextIntent } from '@/app/lib/notebook/context-surface';
 import { useWorkspaceStore } from '@/app/store/workspace-store';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog';
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -76,11 +73,7 @@ import {
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
 import { Input } from '@/components/ui/input';
-import { Switch } from '@/components/ui/switch';
-import { Textarea } from '@/components/ui/textarea';
 import { cn } from '@/lib/utils';
-
-import { EmailHtmlEditor } from './EmailHtmlEditor';
 
 type EmailAccount = {
   id: string;
@@ -99,150 +92,11 @@ type EmailAccount = {
 
 const EMAIL_BACKGROUND_REFRESH_MS = 60_000;
 
-type EmailComposeMode = 'compose' | 'forward' | 'reply' | 'reply-all';
-type EmailComposeAiMode = 'workspace-agent' | 'quick';
-type EmailComposeTone = 'formal' | 'casual' | 'very-casual';
-
-type EmailComposeContextFile = {
-  isImage?: boolean;
-  name?: string;
-  path: string;
-  type?: 'file' | 'directory';
-};
-
-type EmailComposeAgentUsedContext = {
-  path: string;
-  reason?: string;
-};
-
-type EmailComposeAgentToolEvent = {
-  args?: unknown;
-  contextPath?: string;
-  id: string;
-  label?: string;
-  resultPreview?: string;
-  status: 'running' | 'done';
-  toolName: string;
-};
-
-type EmailComposeDraft = {
-  aiGenerated?: boolean;
-  aiMode: EmailComposeAiMode;
-  aiPrompt: string;
-  aiTone: EmailComposeTone;
-  attachments: EmailAttachmentDraft[];
-  body: string;
-  bodyHtml: string;
-  ccText: string;
-  contextFiles: EmailComposeContextFile[];
-  folder?: string;
-  message?: EmailMessageDetail;
-  mode: EmailComposeMode;
-  subject: string;
-  toText: string;
-  usedContext: EmailComposeAgentUsedContext[];
-};
-
 const MESSAGE_PAGE_SIZE = 20;
-const EMAIL_CONTEXT_FILE_EXTENSIONS = new Set(['txt', 'md', 'markdown', 'csv', 'json', 'pdf']);
 type EmailClientProps = {
   contextIntent?: NotebookEmailContextIntent | null;
   embedded?: boolean;
 };
-
-function splitRecipientInput(value: string): string[] {
-  return value
-    .split(/[,\n;]/u)
-    .map((entry) => extractEmailAddressForCompose(entry) || entry.trim())
-    .filter(Boolean);
-}
-
-function isValidComposeRecipient(value: string): boolean {
-  return /^[^\s@<>",;]+@[^\s@<>",;]+\.[^\s@<>",;]+$/u.test(value.trim());
-}
-
-function normalizeComposeRecipient(value: string): string {
-  return extractEmailAddressForCompose(value) || value.trim();
-}
-
-function visibleEmailAttachments(attachments: EmailAttachmentDraft[]): EmailAttachmentDraft[] {
-  return attachments.filter((attachment) => attachment.disposition !== 'inline');
-}
-
-function mergeVisibleEmailAttachments(current: EmailAttachmentDraft[], visible: EmailAttachmentDraft[]): EmailAttachmentDraft[] {
-  return [
-    ...current.filter((attachment) => attachment.disposition === 'inline'),
-    ...visible,
-  ];
-}
-
-function referencedInlineContentIds(html: string): Set<string> {
-  const ids = new Set<string>();
-  const imagePattern = /<img\b[^>]*\ssrc\s*=\s*("cid:([^"]+)"|'cid:([^']+)'|cid:([^\s"'>]+))/giu;
-  let match: RegExpExecArray | null;
-
-  while ((match = imagePattern.exec(html)) !== null) {
-    const contentId = (match[2] || match[3] || match[4] || '').trim();
-    if (contentId) ids.add(contentId);
-  }
-
-  return ids;
-}
-
-function pruneUnreferencedInlineEmailAttachments(attachments: EmailAttachmentDraft[], html: string): EmailAttachmentDraft[] {
-  const referencedIds = referencedInlineContentIds(html);
-  return attachments.filter((attachment) => (
-    attachment.disposition !== 'inline'
-    || (attachment.contentId && referencedIds.has(attachment.contentId))
-  ));
-}
-
-function appendComposeRecipients(current: string[], additions: string[]): string[] {
-  const seen = new Set(current.map((recipient) => recipient.trim().toLowerCase()).filter(Boolean));
-  const next = [...current];
-  for (const addition of additions) {
-    const recipient = normalizeComposeRecipient(addition);
-    const key = recipient.toLowerCase();
-    if (!recipient || seen.has(key)) continue;
-    seen.add(key);
-    next.push(recipient);
-  }
-  return next;
-}
-
-function extractRecipientEmailsForCompose(value: string[] | string | undefined): string[] {
-  if (!value) return [];
-  if (Array.isArray(value)) return value.flatMap(extractRecipientEmailsForCompose);
-  return splitRecipientInput(value);
-}
-
-function uniqueComposeRecipients(values: string[], ownAddresses = new Set<string>()): string[] {
-  const seen = new Set<string>();
-  const recipients: string[] = [];
-  for (const value of values) {
-    const email = extractEmailAddressForCompose(value);
-    if (!email || ownAddresses.has(email) || seen.has(email)) continue;
-    seen.add(email);
-    recipients.push(email);
-  }
-  return recipients;
-}
-
-function composeRecipientText(values: string[]) {
-  return values.join(', ');
-}
-
-function replySubjectForCompose(subject: string) {
-  const normalized = subject.trim();
-  if (!normalized) return 'Re:';
-  return /^re:/iu.test(normalized) ? normalized : `Re: ${normalized}`;
-}
-
-function forwardSubjectForCompose(subject: string) {
-  const normalized = subject.trim();
-  if (!normalized) return 'Fwd:';
-  return /^(fwd|fw):/iu.test(normalized) ? normalized : `Fwd: ${normalized}`;
-}
 
 function isFetchNetworkError(error: unknown): boolean {
   return error instanceof TypeError && /failed to fetch|fetch failed|networkerror/iu.test(error.message);
@@ -423,770 +277,6 @@ async function readEmailAiDraftStream(
 
   if (buffer.trim()) processEvent(buffer);
   return body;
-}
-
-function fileExtension(filePath: string): string {
-  return filePath.split('.').pop()?.toLowerCase() || '';
-}
-
-function isSupportedEmailContextFile(file: FilePickerFile | EmailComposeContextFile): boolean {
-  return file.type !== 'directory' && EMAIL_CONTEXT_FILE_EXTENSIONS.has(fileExtension(file.path || file.name || ''));
-}
-
-function contextFileName(file: EmailComposeContextFile): string {
-  return file.name || file.path.split('/').pop() || file.path;
-}
-
-function normalizeAgentUsedContext(value: unknown): EmailComposeAgentUsedContext[] {
-  if (!Array.isArray(value)) return [];
-  const seen = new Set<string>();
-  const output: EmailComposeAgentUsedContext[] = [];
-  for (const entry of value) {
-    if (!entry || typeof entry !== 'object' || Array.isArray(entry)) continue;
-    const record = entry as Record<string, unknown>;
-    const path = String(record.path || '').trim();
-    if (!path || seen.has(path)) continue;
-    seen.add(path);
-    const reason = String(record.reason || '').trim();
-    output.push(reason ? { path, reason } : { path });
-  }
-  return output;
-}
-
-function formatToolPreview(value: unknown): string {
-  if (!value || typeof value !== 'object') return '';
-  return Object.entries(value as Record<string, unknown>)
-    .slice(0, 4)
-    .map(([key, entry]) => `${key}: ${String(entry).slice(0, 80)}`)
-    .join(', ');
-}
-
-type EmailComposeDialogLabels = Pick<EmailMessageViewerLabels, 'cc' | 'date' | 'emptyBody' | 'from' | 'noSubject' | 'remoteImagesBlocked' | 'showRemoteImages' | 'to'> & {
-  attachmentsAdd: string;
-  attachmentsAllFiles: string;
-  attachmentsAttached: string;
-  attachmentsCancel: string;
-  attachmentsConfirm: string;
-  attachmentsDialogDescription: string;
-  attachmentsDialogTitle: string;
-  attachmentsEmpty: string;
-  attachmentsLimitExceeded: string;
-  attachmentsLoading: string;
-  attachmentsFolders: string;
-  attachmentsRefresh: string;
-  attachmentsRemove: string;
-  attachmentsSearchPlaceholder: string;
-  attachmentsSortBy: string;
-  attachmentsSortCreated: string;
-  attachmentsSortModified: string;
-  attachmentsSortName: string;
-  attachmentsSortSize: string;
-  attachmentsSelectFiles: string;
-  attachmentsSendMarkdownAsPdf: string;
-  attachmentsSendMarkdownAsPdfShort: string;
-  attachmentsTabUpload: string;
-  attachmentsTabWorkspace: string;
-  attachmentsUploadDrop: string;
-  attachmentsUploadHint: string;
-  attachmentsUsageLabel: string;
-  cancel: string;
-  composeAiReplyTitle: string;
-  composeAiPromptLabel: string;
-  composeAiPromptPlaceholder: string;
-  composeBodyLabel: string;
-  composeBodyPlaceholder: string;
-  composeDescription: string;
-  composeForwardTitle: string;
-  composeAddContext: string;
-  composeAgentReady: string;
-  composeAgentToolDetails: string;
-  composeAgentWorking: string;
-  composeAiDraftReady: string;
-  composeAiReadingContext: string;
-  composeAiModeQuick: string;
-  composeAiModeWorkspaceAgent: string;
-  composeAiWritingDraft: string;
-  composeGenerateWithAi: string;
-  composeGeneratingWithAi: string;
-  composeContextFiles: string;
-  composeNoContextFiles: string;
-  composeNewTitle: string;
-  composeWorkspaceOutboxDescription: string;
-  composeWorkspaceOutboxTitle: string;
-  composeOriginalTitle: string;
-  composeReferencePickerEmpty: string;
-  composeReferencePickerHeader: string;
-  composeReferencePickerSearchPlaceholder: string;
-  composeRemoveContextFile: string;
-  composeReplyAllTitle: string;
-  composeReplyTitle: string;
-  composeSaveDraft: string;
-  composeSavingDraft: string;
-  composeDraftSaved: string;
-  composeSend: string;
-  composeSending: string;
-  composeToneCasual: string;
-  composeToneFormal: string;
-  composeToneLabel: string;
-  composeToneVeryCasual: string;
-  composeUsedContext: string;
-  subject: string;
-};
-
-function composeDialogTitle(draft: EmailComposeDraft, labels: EmailComposeDialogLabels) {
-  if (draft.mode === 'compose') return labels.composeNewTitle;
-  if (draft.aiGenerated) return labels.composeAiReplyTitle;
-  if (draft.mode === 'forward') return labels.composeForwardTitle;
-  if (draft.mode === 'reply-all') return labels.composeReplyAllTitle;
-  return labels.composeReplyTitle;
-}
-
-function EmailComposeAgentProgress({
-  events,
-  labels,
-  locale,
-  status,
-  usedContext,
-}: {
-  events: EmailComposeAgentToolEvent[];
-  labels: EmailComposeDialogLabels;
-  locale: string;
-  status: string | null;
-  usedContext: EmailComposeAgentUsedContext[];
-}) {
-  if (!status && events.length === 0 && usedContext.length === 0) return null;
-
-  return (
-    <div className="space-y-2 border border-border bg-background px-3 py-2 text-xs">
-      <div className="flex items-center justify-between gap-2">
-        <div className="flex min-w-0 items-center gap-2 font-medium">
-          <Sparkles className="h-3.5 w-3.5 shrink-0 text-primary" />
-          <span className="min-w-0 truncate">{status || labels.composeAgentReady}</span>
-        </div>
-        {events.some((event) => event.status === 'running') ? (
-          <Loader2 className="h-3.5 w-3.5 shrink-0 animate-spin text-muted-foreground" />
-        ) : null}
-      </div>
-      {events.length > 0 ? (
-        <div className="space-y-1.5">
-          {events.map((event) => {
-            const display = getToolDisplayInfo(event.toolName, locale);
-            const preview = event.resultPreview || event.contextPath || formatToolPreview(event.args);
-            return (
-              <div key={event.id} className="flex items-start gap-2 border border-border/70 bg-muted/35 px-2 py-1.5">
-                <div className="mt-0.5 flex h-5 w-5 shrink-0 items-center justify-center border border-border bg-background">
-                  {event.status === 'running' ? (
-                    <Loader2 className="h-3 w-3 animate-spin text-muted-foreground" />
-                  ) : (
-                    <Wrench className="h-3 w-3 text-muted-foreground" />
-                  )}
-                </div>
-                <div className="min-w-0">
-                  <div className="font-medium text-foreground">{event.label || display.label}</div>
-                  {preview ? (
-                    <div className="mt-0.5 line-clamp-2 break-words text-muted-foreground" title={preview}>
-                      {preview}
-                    </div>
-                  ) : null}
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      ) : null}
-      {usedContext.length > 0 ? (
-        <div className="border-t border-border pt-2">
-          <div className="mb-1 font-medium text-muted-foreground">{labels.composeUsedContext}</div>
-          <div className="flex flex-wrap gap-1.5">
-            {usedContext.map((entry) => (
-              <span
-                key={entry.path}
-                className="inline-flex max-w-full items-center gap-1 border border-border bg-muted/40 px-2 py-1"
-                title={entry.reason || entry.path}
-              >
-                <FileText className="h-3 w-3 shrink-0 text-muted-foreground" />
-                <span className="min-w-0 truncate">{entry.path}</span>
-              </span>
-            ))}
-          </div>
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function EmailRecipientChipInput({
-  disabled,
-  id,
-  onChange,
-  value,
-}: {
-  disabled?: boolean;
-  id: string;
-  onChange: (value: string) => void;
-  value: string;
-}) {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [draftValue, setDraftValue] = useState('');
-  const recipients = useMemo(() => splitRecipientInput(value), [value]);
-
-  const setRecipients = useCallback((nextRecipients: string[]) => {
-    onChange(composeRecipientText(nextRecipients));
-  }, [onChange]);
-
-  const commitRecipients = useCallback((rawValue = draftValue) => {
-    const additions = splitRecipientInput(rawValue);
-    if (additions.length === 0) return false;
-    setRecipients(appendComposeRecipients(recipients, additions));
-    setDraftValue('');
-    return true;
-  }, [draftValue, recipients, setRecipients]);
-
-  const handleDraftChange = useCallback((event: ChangeEvent<HTMLInputElement>) => {
-    const nextValue = event.target.value;
-    if (!/[,\n;]/u.test(nextValue)) {
-      setDraftValue(nextValue);
-      return;
-    }
-
-    const hasTrailingDelimiter = /[,\n;]\s*$/u.test(nextValue);
-    const parts = nextValue.split(/[,\n;]/u);
-    const pendingValue = hasTrailingDelimiter ? '' : parts.pop() || '';
-    const additions = parts.map(normalizeComposeRecipient).filter(Boolean);
-    if (additions.length > 0) {
-      setRecipients(appendComposeRecipients(recipients, additions));
-    }
-    setDraftValue(pendingValue);
-  }, [recipients, setRecipients]);
-
-  const handleKeyDown = useCallback((event: KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === 'Enter' || event.key === 'Tab' || event.key === ',' || event.key === ';') {
-      if (draftValue.trim()) {
-        event.preventDefault();
-        commitRecipients();
-      }
-      return;
-    }
-
-    if (event.key === 'Backspace' && !draftValue && recipients.length > 0) {
-      event.preventDefault();
-      setRecipients(recipients.slice(0, -1));
-    }
-  }, [commitRecipients, draftValue, recipients, setRecipients]);
-
-  const removeRecipient = useCallback((index: number) => {
-    setRecipients(recipients.filter((_, recipientIndex) => recipientIndex !== index));
-    window.setTimeout(() => inputRef.current?.focus(), 0);
-  }, [recipients, setRecipients]);
-
-  return (
-    <div
-      className={cn(
-        'flex min-h-10 w-full flex-wrap items-center gap-1 border border-input bg-background px-2 py-1.5 text-sm focus-within:ring-1 focus-within:ring-ring',
-        disabled && 'opacity-50',
-      )}
-      onClick={() => inputRef.current?.focus()}
-    >
-      {recipients.map((recipient, index) => {
-        const isValid = isValidComposeRecipient(recipient);
-        return (
-          <span
-            key={`${recipient}:${index}`}
-            className={cn(
-              'inline-flex max-w-full items-center gap-1 border bg-muted/40 px-2 py-1 text-xs',
-              isValid ? 'border-border text-foreground' : 'border-destructive/60 bg-destructive/10 text-destructive',
-            )}
-            aria-invalid={!isValid}
-            title={recipient}
-          >
-            <span className="min-w-0 truncate">{recipient}</span>
-            <button
-              type="button"
-              className="shrink-0 text-muted-foreground hover:text-foreground disabled:pointer-events-none"
-              aria-label={`Remove ${recipient}`}
-              onClick={(event) => {
-                event.stopPropagation();
-                removeRecipient(index);
-              }}
-              disabled={disabled}
-            >
-              <X className="h-3 w-3" />
-            </button>
-          </span>
-        );
-      })}
-      <input
-        id={id}
-        ref={inputRef}
-        value={draftValue}
-        onBlur={() => {
-          if (draftValue.trim()) commitRecipients();
-        }}
-        onChange={handleDraftChange}
-        onKeyDown={handleKeyDown}
-        placeholder={recipients.length === 0 ? 'email@example.com' : ''}
-        disabled={disabled}
-        className="min-w-[11rem] flex-1 bg-transparent py-1 text-sm outline-none placeholder:text-muted-foreground disabled:cursor-not-allowed"
-      />
-    </div>
-  );
-}
-
-function EmailComposeDialog({
-  agentEvents,
-  agentStatus,
-  allowRemoteResourcesByDefault,
-  allowedRemoteResourceSenders,
-  draft,
-  error,
-  isGeneratingAi,
-  isSubmitting,
-  isWorkspaceOutboxReview,
-  reviewCase,
-  labels,
-  locale,
-  onAllowRemoteResourcesForSender,
-  onClose,
-  onGenerateAi,
-  onSave,
-  onSubmit,
-  onUpdate,
-}: {
-  agentEvents: EmailComposeAgentToolEvent[];
-  agentStatus: string | null;
-  allowRemoteResourcesByDefault: boolean;
-  allowedRemoteResourceSenders: string[];
-  draft: EmailComposeDraft | null;
-  error: string | null;
-  isGeneratingAi: boolean;
-  isSubmitting: boolean;
-  isWorkspaceOutboxReview: boolean;
-  reviewCase: WorkspaceInboxCase | null;
-  labels: EmailComposeDialogLabels;
-  locale: string;
-  onAllowRemoteResourcesForSender(sender: string): void;
-  onClose(): void;
-  onGenerateAi(): void;
-  onSave(): void;
-  onSubmit(): void;
-  onUpdate(updates: Partial<Pick<EmailComposeDraft, 'aiMode' | 'aiPrompt' | 'aiTone' | 'attachments' | 'body' | 'bodyHtml' | 'ccText' | 'contextFiles' | 'subject' | 'toText' | 'usedContext'>>): void;
-}) {
-  const activeWorkspaceId = useWorkspaceStore((state) => state.activeWorkspaceId);
-  const [isReferencePickerOpen, setIsReferencePickerOpen] = useState(false);
-  const [activeReferenceMatch, setActiveReferenceMatch] = useState<ComposerReferenceMatch | null>(null);
-  const [referencePickerItems, setReferencePickerItems] = useState<ComposerReferencePickerItem<FilePickerFile>[]>([]);
-  const [referenceSearchQuery, setReferenceSearchQuery] = useState('');
-  const [isReferencePickerLoading, setIsReferencePickerLoading] = useState(false);
-  const [selectedReferenceIndex, setSelectedReferenceIndex] = useState(0);
-  const aiPromptTextareaRef = useRef<HTMLTextAreaElement>(null);
-  const referenceSearchInputRef = useRef<HTMLInputElement>(null);
-  const referenceRequestIdRef = useRef(0);
-  const selectedContextPaths = useMemo(() => new Set((draft?.contextFiles || []).map((file) => file.path)), [draft?.contextFiles]);
-  const displayedAttachments = useMemo(() => visibleEmailAttachments(draft?.attachments || []), [draft?.attachments]);
-  const updateDisplayedAttachments = useCallback((attachments: EmailAttachmentDraft[]) => {
-    if (!draft) return;
-    onUpdate({ attachments: mergeVisibleEmailAttachments(draft.attachments, attachments) });
-  }, [draft, onUpdate]);
-
-  const closeReferencePicker = useCallback(() => {
-    referenceRequestIdRef.current += 1;
-    setIsReferencePickerOpen(false);
-    setActiveReferenceMatch(null);
-    setReferencePickerItems([]);
-    setReferenceSearchQuery('');
-    setSelectedReferenceIndex(0);
-    setIsReferencePickerLoading(false);
-  }, []);
-
-  const loadReferenceFiles = useCallback(async (query = '') => {
-    const requestId = referenceRequestIdRef.current + 1;
-    referenceRequestIdRef.current = requestId;
-    setIsReferencePickerLoading(true);
-    try {
-      if (!activeWorkspaceId) throw new Error('Workspace context is not ready');
-      const files = await listWorkspaceFileReferences({ query, limit: 50, workspaceId: activeWorkspaceId });
-      if (requestId !== referenceRequestIdRef.current) return;
-      const items = files
-        .filter(isSupportedEmailContextFile)
-        .filter((file) => !selectedContextPaths.has(file.path))
-        .map((file) => ({
-          id: `file:${file.path}`,
-          kind: 'file' as const,
-          icon: getFileIconComponent({ name: file.name, path: file.path, type: file.type }),
-          label: file.path,
-          payload: file,
-          secondaryLabel: file.name,
-        }));
-      setReferencePickerItems(items);
-      setSelectedReferenceIndex(0);
-    } catch {
-      if (requestId !== referenceRequestIdRef.current) return;
-      setReferencePickerItems([]);
-      setSelectedReferenceIndex(0);
-    } finally {
-      if (requestId === referenceRequestIdRef.current) {
-        setIsReferencePickerLoading(false);
-      }
-    }
-  }, [activeWorkspaceId, selectedContextPaths]);
-
-  const selectReferenceFile = useCallback((item: ComposerReferencePickerItem<FilePickerFile>) => {
-    const file = item.payload;
-    if (!draft || !isSupportedEmailContextFile(file)) return;
-    const currentAiPrompt = aiPromptTextareaRef.current?.value ?? draft.aiPrompt;
-    let nextAiPrompt = currentAiPrompt;
-    let nextCursorPosition: number | null = null;
-
-    if (activeReferenceMatch) {
-      const replacement = `+"${file.path}" `;
-      const nextPrompt = replaceComposerReference(currentAiPrompt, activeReferenceMatch, replacement);
-      nextAiPrompt = nextPrompt.nextValue;
-      nextCursorPosition = nextPrompt.nextCursorPosition;
-    }
-
-    onUpdate({
-      aiPrompt: nextAiPrompt,
-      contextFiles: selectedContextPaths.has(file.path)
-        ? draft.contextFiles
-        : [
-            ...draft.contextFiles,
-            {
-              isImage: file.isImage,
-              name: file.name,
-              path: file.path,
-              type: file.type,
-            },
-          ],
-      usedContext: [],
-    });
-    closeReferencePicker();
-
-    if (nextCursorPosition !== null) {
-      window.setTimeout(() => {
-        aiPromptTextareaRef.current?.focus();
-        aiPromptTextareaRef.current?.setSelectionRange(nextCursorPosition, nextCursorPosition);
-      }, 0);
-    }
-  }, [activeReferenceMatch, closeReferencePicker, draft, onUpdate, selectedContextPaths]);
-
-  const openManualReferencePicker = useCallback(() => {
-    const nextOpen = !isReferencePickerOpen;
-    if (!nextOpen) {
-      closeReferencePicker();
-      return;
-    }
-    setActiveReferenceMatch(null);
-    setReferenceSearchQuery('');
-    setIsReferencePickerOpen(true);
-    void loadReferenceFiles('');
-    window.setTimeout(() => referenceSearchInputRef.current?.focus(), 0);
-  }, [closeReferencePicker, isReferencePickerOpen, loadReferenceFiles]);
-
-  const updateReferenceSearch = useCallback((value: string) => {
-    setReferenceSearchQuery(value);
-    void loadReferenceFiles(value);
-  }, [loadReferenceFiles]);
-
-  const handleReferenceSearchKeyDown = useCallback((event: KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === 'ArrowDown') {
-      event.preventDefault();
-      setSelectedReferenceIndex((current) => Math.min(referencePickerItems.length - 1, current + 1));
-      return;
-    }
-    if (event.key === 'ArrowUp') {
-      event.preventDefault();
-      setSelectedReferenceIndex((current) => Math.max(0, current - 1));
-      return;
-    }
-    if (event.key === 'Enter') {
-      event.preventDefault();
-      const item = referencePickerItems[selectedReferenceIndex];
-      if (item) selectReferenceFile(item);
-      return;
-    }
-    if (event.key === 'Escape') {
-      event.preventDefault();
-      closeReferencePicker();
-    }
-  }, [closeReferencePicker, referencePickerItems, selectReferenceFile, selectedReferenceIndex]);
-
-  const handleAiPromptChange = useCallback((event: ChangeEvent<HTMLTextAreaElement>) => {
-    const value = event.target.value;
-    const cursorPosition = event.target.selectionStart;
-    onUpdate({ aiPrompt: value });
-
-    if (!draft || draft.aiMode !== 'workspace-agent') {
-      closeReferencePicker();
-      return;
-    }
-
-    const match = findActiveComposerReference(value, cursorPosition);
-    if (match?.kind === 'file' && match.trigger === '+') {
-      setActiveReferenceMatch(match);
-      setReferenceSearchQuery(match.query);
-      setIsReferencePickerOpen(true);
-      void loadReferenceFiles(match.query);
-      return;
-    }
-
-    if (activeReferenceMatch) {
-      closeReferencePicker();
-    }
-  }, [activeReferenceMatch, closeReferencePicker, draft, loadReferenceFiles, onUpdate]);
-
-  return (
-    <Dialog
-      open={Boolean(draft)}
-      onOpenChange={(open) => {
-        if (!open && !isSubmitting && !isGeneratingAi) onClose();
-      }}
-    >
-      <DialogContent layout="viewport">
-        {draft && (
-          <>
-            <DialogHeader className="shrink-0 border-b border-border px-4 py-3 pr-10 sm:px-5">
-              <DialogTitle className="text-base leading-6">{isWorkspaceOutboxReview ? labels.composeWorkspaceOutboxTitle : composeDialogTitle(draft, labels)}</DialogTitle>
-              <DialogDescription className="text-xs leading-5 sm:text-sm">{isWorkspaceOutboxReview ? labels.composeWorkspaceOutboxDescription : labels.composeDescription}</DialogDescription>
-            </DialogHeader>
-            <div className="min-h-0 flex-1 overflow-y-auto px-3 py-3 sm:px-5">
-              {isWorkspaceOutboxReview && reviewCase ? (
-                <section className="mb-3 rounded-lg border border-primary/15 bg-primary/[0.035] px-3 py-2.5 text-sm">
-                  <div className="flex flex-wrap items-center gap-2"><span className="text-xs font-semibold uppercase tracking-[0.12em] text-primary">{labels.composeOriginalTitle}</span><Badge variant="outline">{reviewCase.priority}</Badge><Badge variant="secondary">{reviewCase.status}</Badge></div>
-                  <p className="mt-1.5 truncate font-medium">{reviewCase.requesterName || reviewCase.requesterAddress || reviewCase.subject}</p>
-                  <p className="mt-0.5 truncate text-xs text-muted-foreground">{reviewCase.subject}</p>
-                </section>
-              ) : null}
-              <div className={cn('grid min-h-full gap-3', draft.message && 'lg:grid-cols-[minmax(300px,420px)_minmax(0,1fr)]')}>
-                <section className="min-w-0 space-y-3">
-                  <div className="space-y-1.5">
-                    <label className="text-xs font-semibold uppercase tracking-[0.14em] text-muted-foreground" htmlFor="email-compose-to">
-                      {labels.to}
-                    </label>
-                    <EmailRecipientChipInput
-                      id="email-compose-to"
-                      value={draft.toText}
-                      onChange={(value) => onUpdate({ toText: value })}
-                      disabled={isSubmitting}
-                    />
-                  </div>
-                  <div className="space-y-1.5">
-                    <label className="text-xs font-semibold uppercase tracking-[0.14em] text-muted-foreground" htmlFor="email-compose-cc">
-                      {labels.cc}
-                    </label>
-                    <EmailRecipientChipInput
-                      id="email-compose-cc"
-                      value={draft.ccText}
-                      onChange={(value) => onUpdate({ ccText: value })}
-                      disabled={isSubmitting}
-                    />
-                  </div>
-                  <div className="space-y-1.5">
-                    <label className="text-xs font-semibold uppercase tracking-[0.14em] text-muted-foreground" htmlFor="email-compose-subject">
-                      {labels.subject}
-                    </label>
-                    <Input
-                      id="email-compose-subject"
-                      value={draft.subject}
-                      onChange={(event) => onUpdate({ subject: event.target.value })}
-                      disabled={isSubmitting}
-                    />
-                  </div>
-                  <div className="space-y-2 border border-border bg-muted/30 px-3 py-3">
-                    <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
-                      <label className="text-xs font-semibold uppercase tracking-[0.14em] text-muted-foreground" htmlFor="email-compose-ai-prompt">
-                        {labels.composeAiPromptLabel}
-                      </label>
-                      <div className="flex items-center gap-2 text-xs text-muted-foreground">
-                        <span>{draft.aiMode === 'workspace-agent' ? labels.composeAiModeWorkspaceAgent : labels.composeAiModeQuick}</span>
-                        <Switch
-                          checked={draft.aiMode === 'workspace-agent'}
-                          onCheckedChange={(checked) => onUpdate({
-                            aiMode: checked ? 'workspace-agent' : 'quick',
-                            usedContext: [],
-                          })}
-                          disabled={isSubmitting || isGeneratingAi}
-                          aria-label={labels.composeAiModeWorkspaceAgent}
-                        />
-                      </div>
-                    </div>
-                    <Textarea
-                      id="email-compose-ai-prompt"
-                      ref={aiPromptTextareaRef}
-                      value={draft.aiPrompt}
-                      onChange={handleAiPromptChange}
-                      placeholder={labels.composeAiPromptPlaceholder}
-                      className="min-h-20 resize-y bg-background"
-                      disabled={isSubmitting || isGeneratingAi}
-                    />
-                    <div className="flex flex-col gap-2 sm:flex-row sm:flex-wrap sm:items-center">
-                      <div className="flex flex-col gap-2 sm:flex-row sm:items-center">
-                        <label className="flex items-center gap-2 text-xs text-muted-foreground">
-                          <span>{labels.composeToneLabel}</span>
-                          <select
-                            className="h-8 border border-input bg-background px-2 text-sm text-foreground"
-                            value={draft.aiTone}
-                            onChange={(event) => onUpdate({ aiTone: event.target.value as EmailComposeTone })}
-                            disabled={isSubmitting || isGeneratingAi}
-                          >
-                            <option value="formal">{labels.composeToneFormal}</option>
-                            <option value="casual">{labels.composeToneCasual}</option>
-                            <option value="very-casual">{labels.composeToneVeryCasual}</option>
-                          </select>
-                        </label>
-                        {draft.aiMode === 'workspace-agent' ? (
-                          <div className="relative w-full sm:w-auto">
-                            <Button
-                              type="button"
-                              variant="outline"
-                              size="sm"
-                              className="w-full justify-start sm:w-auto"
-                              onClick={openManualReferencePicker}
-                              disabled={isSubmitting || isGeneratingAi}
-                            >
-                              <Plus className="mr-2 h-4 w-4" />
-                              {labels.composeAddContext}
-                            </Button>
-                            {isReferencePickerOpen ? (
-                              <ComposerReferencePicker
-                                className="min-w-[20rem] max-w-[min(32rem,calc(100vw-3rem))]"
-                                emptyState={labels.composeReferencePickerEmpty}
-                                header={labels.composeReferencePickerHeader}
-                                isLoading={isReferencePickerLoading}
-                                items={referencePickerItems}
-                                onSelect={selectReferenceFile}
-                                onSearchKeyDown={handleReferenceSearchKeyDown}
-                                onSearchValueChange={updateReferenceSearch}
-                                searchAutoFocus={!activeReferenceMatch}
-                                searchInputRef={referenceSearchInputRef}
-                                searchPlaceholder={labels.composeReferencePickerSearchPlaceholder}
-                                searchValue={referenceSearchQuery}
-                                selectedIndex={selectedReferenceIndex}
-                              />
-                            ) : null}
-                          </div>
-                        ) : null}
-                      </div>
-                      <Button
-                        type="button"
-                        variant="outline"
-                        size="sm"
-                        className="w-full sm:w-auto"
-                        onClick={onGenerateAi}
-                        disabled={isSubmitting || isGeneratingAi || !draft.aiPrompt.trim()}
-                      >
-                        {isGeneratingAi ? <Loader2 className="mr-2 h-4 w-4 animate-spin" /> : <Sparkles className="mr-2 h-4 w-4" />}
-                        {isGeneratingAi ? labels.composeGeneratingWithAi : labels.composeGenerateWithAi}
-                      </Button>
-                    </div>
-                    {draft.aiMode === 'workspace-agent' && draft.contextFiles.length > 0 ? (
-                      <div className="space-y-1.5">
-                        <div className="text-xs font-medium text-muted-foreground">{labels.composeContextFiles}</div>
-                        <div className="flex flex-wrap gap-1.5">
-                          {draft.contextFiles.map((file) => (
-                            <span
-                              key={file.path}
-                              className="inline-flex max-w-full items-center gap-1.5 border border-border bg-background px-2 py-1 text-xs"
-                              title={file.path}
-                            >
-                              <FileText className="h-3 w-3 shrink-0 text-muted-foreground" />
-                              <span className="min-w-0 truncate">{contextFileName(file)}</span>
-                              <button
-                                type="button"
-                                className="shrink-0 text-muted-foreground hover:text-foreground"
-                                aria-label={labels.composeRemoveContextFile}
-                                title={labels.composeRemoveContextFile}
-                                onClick={() => onUpdate({
-                                  contextFiles: draft.contextFiles.filter((entry) => entry.path !== file.path),
-                                  usedContext: [],
-                                })}
-                                disabled={isSubmitting || isGeneratingAi}
-                              >
-                                <X className="h-3 w-3" />
-                              </button>
-                            </span>
-                          ))}
-                        </div>
-                      </div>
-                    ) : null}
-                    <EmailComposeAgentProgress
-                      events={agentEvents}
-                      labels={labels}
-                      locale={locale}
-                      status={agentStatus || (isGeneratingAi ? labels.composeAgentWorking : null)}
-                      usedContext={draft.usedContext}
-                    />
-                  </div>
-                  <div className="space-y-1.5">
-                    <label className="text-xs font-semibold uppercase tracking-[0.14em] text-muted-foreground" htmlFor="email-compose-body">
-                      {labels.composeBodyLabel}
-                    </label>
-                    <EmailHtmlEditor
-                      attachments={draft.attachments}
-                      id="email-compose-body"
-                      value={draft.bodyHtml}
-                      onChange={({ html, text }) => onUpdate({ body: text, bodyHtml: html })}
-                      onAttachmentsChange={(attachments) => onUpdate({ attachments })}
-                      placeholder={labels.composeBodyPlaceholder}
-                      disabled={isSubmitting || isGeneratingAi}
-                    />
-                  </div>
-                  <EmailAttachmentPanel
-                    attachments={displayedAttachments}
-                    disabled={isSubmitting || isGeneratingAi}
-                    labels={labels}
-                    onChange={updateDisplayedAttachments}
-                  />
-                  {error && (
-                    <div className="border border-destructive/40 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-                      <p className="break-words">{error}</p>
-                    </div>
-                  )}
-                </section>
-
-                {draft.message && (
-                  <section className="min-w-0 overflow-hidden border border-border bg-card">
-                    <div className="border-b border-border px-3 py-3 sm:px-4">
-                      <div className="text-xs font-semibold uppercase tracking-[0.14em] text-muted-foreground">
-                        {labels.composeOriginalTitle}
-                      </div>
-                      <h3 className="mt-2 truncate text-base font-semibold">{draft.message.subject || labels.noSubject}</h3>
-                      <div className="mt-2 space-y-1 text-xs text-muted-foreground">
-                        <p className="truncate"><span className="font-medium text-foreground">{labels.from}:</span> {draft.message.from}</p>
-                        {formatRecipients(draft.message.to) && <p className="truncate"><span className="font-medium text-foreground">{labels.to}:</span> {formatRecipients(draft.message.to)}</p>}
-                        {formatRecipients(draft.message.cc) && <p className="truncate"><span className="font-medium text-foreground">{labels.cc}:</span> {formatRecipients(draft.message.cc)}</p>}
-                        {draft.message.date && <p className="truncate"><span className="font-medium text-foreground">{labels.date}:</span> {formatDate(draft.message.date)}</p>}
-                      </div>
-                    </div>
-                    <div className="max-h-[42dvh] overflow-y-auto px-3 py-3 sm:px-4 lg:max-h-[calc(100dvh-20rem)]">
-                      <EmailMessageBody
-                        allowRemoteResourcesByDefault={allowRemoteResourcesByDefault}
-                        allowedRemoteResourceSenders={allowedRemoteResourceSenders}
-                        message={draft.message}
-                        onAllowRemoteResourcesForSender={onAllowRemoteResourcesForSender}
-                        emptyText={labels.emptyBody}
-                        remoteImagesBlockedText={labels.remoteImagesBlocked}
-                        showRemoteImagesText={labels.showRemoteImages}
-                      />
-                    </div>
-                  </section>
-                )}
-              </div>
-            </div>
-            <DialogFooter className="shrink-0 border-t border-border px-4 py-3 sm:px-6">
-              <Button type="button" variant="outline" onClick={onClose} disabled={isSubmitting || isGeneratingAi}>
-                {labels.cancel}
-              </Button>
-              {isWorkspaceOutboxReview ? <Button type="button" variant="outline" onClick={onSave} disabled={isSubmitting || isGeneratingAi}>{isSubmitting ? <Loader2 className="mr-2 h-4 w-4 animate-spin" /> : null}{isSubmitting ? labels.composeSavingDraft : labels.composeSaveDraft}</Button> : null}
-              <Button type="button" onClick={onSubmit} disabled={isSubmitting || isGeneratingAi}>
-                {isSubmitting ? <Loader2 className="mr-2 h-4 w-4 animate-spin" /> : <Mail className="mr-2 h-4 w-4" />}
-                {isSubmitting ? labels.composeSending : labels.composeSend}
-              </Button>
-            </DialogFooter>
-          </>
-        )}
-      </DialogContent>
-    </Dialog>
-  );
 }
 
 export function EmailClient({
