+++ app/apps/email/components/email-client-types.ts
@@ -67,6 +67,122 @@ export type EmailMessageDetail = EmailMessageSummary & {
   }>;
 };
 
+export type EmailComposeMode = 'compose' | 'forward' | 'reply' | 'reply-all';
+export type EmailComposeAiMode = 'workspace-agent' | 'quick';
+export type EmailComposeTone = 'formal' | 'casual' | 'very-casual';
+
+export type EmailComposeContextFile = {
+  isImage?: boolean;
+  name?: string;
+  path: string;
+  type?: 'file' | 'directory';
+};
+
+export type EmailComposeAgentUsedContext = {
+  path: string;
+  reason?: string;
+};
+
+export type EmailComposeAgentToolEvent = {
+  args?: unknown;
+  contextPath?: string;
+  id: string;
+  label?: string;
+  resultPreview?: string;
+  status: 'running' | 'done';
+  toolName: string;
+};
+
+export type EmailComposeDraft = {
+  aiGenerated?: boolean;
+  aiMode: EmailComposeAiMode;
+  aiPrompt: string;
+  aiTone: EmailComposeTone;
+  attachments: EmailAttachmentDraft[];
+  body: string;
+  bodyHtml: string;
+  ccText: string;
+  contextFiles: EmailComposeContextFile[];
+  folder?: string;
+  message?: EmailMessageDetail;
+  mode: EmailComposeMode;
+  subject: string;
+  toText: string;
+  usedContext: EmailComposeAgentUsedContext[];
+};
+
+export type EmailComposeDialogLabels = Pick<EmailMessageViewerLabels, 'cc' | 'date' | 'emptyBody' | 'from' | 'noSubject' | 'remoteImagesBlocked' | 'showRemoteImages' | 'to'> & {
+  attachmentsAdd: string;
+  attachmentsAllFiles: string;
+  attachmentsAttached: string;
+  attachmentsCancel: string;
+  attachmentsConfirm: string;
+  attachmentsDialogDescription: string;
+  attachmentsDialogTitle: string;
+  attachmentsEmpty: string;
+  attachmentsLimitExceeded: string;
+  attachmentsLoading: string;
+  attachmentsFolders: string;
+  attachmentsRefresh: string;
+  attachmentsRemove: string;
+  attachmentsSearchPlaceholder: string;
+  attachmentsSortBy: string;
+  attachmentsSortCreated: string;
+  attachmentsSortModified: string;
+  attachmentsSortName: string;
+  attachmentsSortSize: string;
+  attachmentsSelectFiles: string;
+  attachmentsSendMarkdownAsPdf: string;
+  attachmentsSendMarkdownAsPdfShort: string;
+  attachmentsTabUpload: string;
+  attachmentsTabWorkspace: string;
+  attachmentsUploadDrop: string;
+  attachmentsUploadHint: string;
+  attachmentsUsageLabel: string;
+  cancel: string;
+  composeAiReplyTitle: string;
+  composeAiPromptLabel: string;
+  composeAiPromptPlaceholder: string;
+  composeBodyLabel: string;
+  composeBodyPlaceholder: string;
+  composeDescription: string;
+  composeForwardTitle: string;
+  composeAddContext: string;
+  composeAgentReady: string;
+  composeAgentToolDetails: string;
+  composeAgentWorking: string;
+  composeAiDraftReady: string;
+  composeAiReadingContext: string;
+  composeAiModeQuick: string;
+  composeAiModeWorkspaceAgent: string;
+  composeAiWritingDraft: string;
+  composeGenerateWithAi: string;
+  composeGeneratingWithAi: string;
+  composeContextFiles: string;
+  composeNoContextFiles: string;
+  composeNewTitle: string;
+  composeWorkspaceOutboxDescription: string;
+  composeWorkspaceOutboxTitle: string;
+  composeOriginalTitle: string;
+  composeReferencePickerEmpty: string;
+  composeReferencePickerHeader: string;
+  composeReferencePickerSearchPlaceholder: string;
+  composeRemoveContextFile: string;
+  composeReplyAllTitle: string;
+  composeReplyTitle: string;
+  composeSaveDraft: string;
+  composeSavingDraft: string;
+  composeDraftSaved: string;
+  composeSend: string;
+  composeSending: string;
+  composeToneCasual: string;
+  composeToneFormal: string;
+  composeToneLabel: string;
+  composeToneVeryCasual: string;
+  composeUsedContext: string;
+  subject: string;
+};
+
 export type EmailMessageViewerLabels = {
   aiReply: string;
   aiSummary: string;
